requirements.txt
amiyabot==2.0.4
attrdict~=2.0.1
baidu-aip
cos-python-sdk-v5~=1.9.30
dhash~=1.3
fake-useragent==1.5.1
fastapi~=0.79.0
gitpython
jieba~=0.42.1
jinja2~=3.1.2
jionlp~=1.4.19
jsonschema~=4.17.3
lxml_html_clean~=0.1.0
opencv-python~=4.7.0.72
peewee~=3.14.10
pillow==9.5.0
pypinyin~=0.46.0
pyyaml~=6.0
requests-html~=0.10.0
srsly~=2.4.3
starlette~=0.19.1
tencentcloud-sdk-python~=3.0.0
zhon~=1.1.5